extract config for REST
